# Cache Resource and Instrumentationlibrary protos.