Since Resource and IL are essentially constants we can reduce GC pressure by caching conversion to proto. An interesting takeaway is Before actually GC's while After doesn't, I think it demonstrates going from affecting an app's GC to not affecting it for these two codepaths.

After
```
Benchmark                                             Mode  Cnt   Score    Error   Units
CommonAdapterBenchmark.toProto                        avgt   45  11.798 ±  0.055   ns/op
CommonAdapterBenchmark.toProto:·gc.alloc.rate         avgt   45   0.001 ±  0.001  MB/sec
CommonAdapterBenchmark.toProto:·gc.alloc.rate.norm    avgt   45  ≈ 10⁻⁵             B/op
CommonAdapterBenchmark.toProto:·gc.count              avgt   45     ≈ 0           counts
ResourceAdapterBenchmark.toProto                      avgt   45  11.933 ±  0.258   ns/op
ResourceAdapterBenchmark.toProto:·gc.alloc.rate       avgt   45   0.001 ±  0.001  MB/sec
ResourceAdapterBenchmark.toProto:·gc.alloc.rate.norm  avgt   45  ≈ 10⁻⁵             B/op
ResourceAdapterBenchmark.toProto:·gc.count            avgt   45     ≈ 0           counts
```

Before
```
Benchmark                                                      Mode  Cnt     Score    Error   Units
CommonAdapterBenchmark.toProto                                 avgt   45    16.437 ±  0.087   ns/op
CommonAdapterBenchmark.toProto:·gc.alloc.rate                  avgt   45  1544.861 ±  8.070  MB/sec
CommonAdapterBenchmark.toProto:·gc.alloc.rate.norm             avgt   45    40.000 ±  0.001    B/op
CommonAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space         avgt   45  1549.997 ± 36.339  MB/sec
CommonAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm    avgt   45    40.130 ±  0.885    B/op
CommonAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen            avgt   45     0.002 ±  0.001  MB/sec
CommonAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm       avgt   45    ≈ 10⁻⁴             B/op
CommonAdapterBenchmark.toProto:·gc.count                       avgt   45   423.000           counts
CommonAdapterBenchmark.toProto:·gc.time                        avgt   45   184.000               ms
ResourceAdapterBenchmark.toProto                               avgt   45   361.830 ±  8.614   ns/op
ResourceAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  2980.543 ± 52.396  MB/sec
ResourceAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  1696.000 ± 12.518    B/op
ResourceAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  2987.938 ± 73.829  MB/sec
ResourceAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  1700.379 ± 34.620    B/op
ResourceAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.003 ±  0.001  MB/sec
ResourceAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.002 ±  0.001    B/op
ResourceAdapterBenchmark.toProto:·gc.count                     avgt   45   624.000           counts
ResourceAdapterBenchmark.toProto:·gc.time                      avgt   45   307.000               ms
```